python_workers/jobspy_fetcher/requirements.txt
python-jobspy>=1.1.70
boto3>=1.34.0
requests>=2.31.0

